Get confirmation of email working
There was nothing wrong with the iframe and accessing content in it. The issue was that the focus remained on the previous tab so the `expect()` would always fail.

Now we record the details of the `window` opened when we click the confirmation link in the email, and we perform our `expect()` calls within the context of it.