Add change account email feature
This change adds the feature and related steps to support testing the change account functionality.

At the time of adding we know that if you change the account on a registration for a user that has multiple registrations, it will work for that one but you then won't be able to change it for the others.