Add new rake task and docs to reset vagrant db
As a first step to being able to create a rake task that can cycle through a range of tests automatically we needed a way of resetting the database between each run. The [waste-carriers-renewals](https://github.com/DEFRA/waste-carriers-renewals) project already includes one to perform the reset. The acceptance tests just needed a way to call it on the Vagrant box.

So this change adds a new rake task called `reset` which calls the command on the Vagrant box using ssh. It also updates the README to provide details about it and what configuration is needed to support.