Import regs from json fixtures dynamically
The previous version of the db reset task specified which json files to import and where to find them.

This might be unknown, and would need to be updated everytime a new one was added. Hence this change which attempts to make the identifying then insert of json fixture files dynamic based on whatever is in the folder.