Standardise environment variables
https://eaflood.atlassian.net/browse/WC-281

To remove any confusion and help standardise usage of environment variables across Waste Carriers the changes here are to bring the project inline with the other apps, and to help remove duplication where possible.

Specifically we rename the one environment variable used  by this project to follow the standard naming convention for the Waste Carriers Service project.