Update frontend LT charity to be pending
We cannot currently complete the lower tier journey because we need to confirm the email. The current code does contain stuff to do this both in mailinator and mailcatcher, but its our intention to add the mail helper we used in WEX going forward.

It shouldn't be too much work to implement, so rather than invest effort getting that old code working, we'll mark anything email related as pending.