Fix reset dbs rake task
The Vagrant box adds the file `/etc/profile.d/myvars.sh` whenever it starts up. This is a bash file of environment variables that should be loaded whenever the box is started using `vagrant up`.

On this basis we assumed when the `reset_dbs` rake task was being run it would see these env vars and so work as if we ran the file whilst ssh'd in using `vagrant ssh`.

Turns out this is not the guess (probably one of those login shell vs non-login shell vs interactive shell things I can never keep on top of).

This changes fixes the issue and ensure the env vars are sourced before we then run any rake tasks in the apps themselves.