Update last-email path
The `/last-email` path is a feature in the WCR frontend which when accessed will return details of the last email sent by an app as JSON.

This is such a useful feature for acceptance testing that we wanted to include it in all our services. So rather than simply copy and paste the same code, we created [defra-ruby-email](https://github.com/DEFRA/defra-ruby-email). It is a rails engine which when mounted into an app and enabled, handles capturing and returning details of the last email sent.

The WCR frontend app has been updated to use defra-ruby-email but this has meant a slight change in the path to `/last-email`.

This change is just about updating the tests to match.