Rename existing StartPage to OldStartPage (#259)
Since we now have a new start page we need to free the name of the page object so that we can define a new object with that name which is what we are going to keep in the long run. This existing StartPage has been renamed as we can delete it as soon as we deploy the new functionality.