Implement new registration lower tier feature
This implements a new feature for lower tier.
As part of this:
 - A new gem has been added to the project that add some useful methods to the standard Ruby library (check https://github.com/rubyworks/facets)
 - New helper method to go through smart answers with random choices

The feature includes two scenarios, one with a Charity and one with a SoleTrader with answers to smart answers

Questions: Shall we add a scenario for completing lower tier after answering "I know I need a lower tier" In the check your tier page?