Change selectors in location page to apply to old and new version of the page
We recently made some fixes to the location page class which was defined more than once. When doing so, we kept one of the 3 versions without realising that it was containing selectors that were specific to the new version of the location page but not the old. This fixes the selectors to use `input[value=blah]` which will apply to both versions of the page and fix the tests