Add support to fetch finance urls directly
This implements a new helper that will store the ID of a registration in an object and will use it to visit finance pages directly.
It is implemented so that it can be called multiple times with the same registration id but will only visit and fetch the data once, then store and play those back to the user via cache class instance objects.