Refactor of finance admin feature (#283)
* Refactor of finance admin feature

This implement more steps using the direct visit of finance page. It also introduce a new step to tell the test suite that a renewal is completed and now operations have to happen on a registration. It also fixes some bugs introduced in the previous PR to deal with fetching registration VS renewal ids from the API