Remove step that used old seeds and refactor tests
This removes a scenario which was expecting existing registration to be seedede in the database.
This will make sure we can run these tests more than once without resetting the database.

An extension to the SeedData class has also been made to use options more smartly and do small twicks to the data. This allowed us to have 1 json per type of business and control account email and business names as required.