Extend SeedData so that seeded data can be retrieved
I was requested by Andrew to see if there is a way we can save the seeded data in a variable so that later in the tests we can access the seeded data and use them to check contents match the info in the seeded registration. This PR implement that change.