Fix payment feature (#305)
* Fix payment feature

Close: https://github.com/DEFRA/waste-carriers-acceptance-tests/issues/300

This PR contains the addition of an helper method that will allow the suite to check if it is running in an environment were mocking is enabled.

This method is then used in a couple of tests that are specific to test worldpay behaviour. In order to skip an entire scenario running, I have created a new `Given` that will appear pending and will output an useful message when mocking is enabled. The first `Given` being pending will cause the scenario to not run.

This also fixes an issue with checking the content of the email when a renewal is submitted with a pending bank transfer. Since we recently updated the content of that email to be in line with what is sent via new registration, we also needed to update the scenario to reflect the latest changes. I chose to use the general email scenario as it was very handy.