Set carrier type after seeding (#307)
We recently made a twick https://github.com/DEFRA/waste-carriers-acceptance-tests/pull/295 so that when we seed data we also keep a variable holding the registration data from which that registration was created. Hence, in the method which require a @carrier to be setted, when it is nil and we have the registration data from seeding, it will use the carrier type that was setted  through the seeding.