Update Travis config (#321)
This updates the travis config to be more consistent with our other ruby based projects.

This includes remove the 'magic' in the Rakefile which handled running on the Travis CI servers. Now its clearer what `bundle exec rake` will do as there is no magic going on behind the scenes.