Move config to its own folder for consistency (#322)
In the [waste-exemptions-acceptance-tests](https://github.com/DEFRA/waste-exemptions-acceptance-tests) the config files are stored in their own `config` folder and have a consistent format for the file names.

Though not critical, it moves a large volume of files that are actually rarely touched out of the root of the project. This just makes things a little easier to navigate in the terminal, in your editor, and in the GitHub UI. It should also help (a little!) remove any belief that the solution is overly complex.

Finally, it moves this project another step closer to being consistent with the WEX acceptance tests, which will help when having to jump in and out of them.