Fix expired registration seeding
Based on my testing I could find no evidence that `SeedData` was doing anything with the `custom_expires` argument. So because that was not getting applied to the seed data, we were not sending it to the back-office API.

It's logic depends on that being set to then apply the custom expires_on date provided.

However, once I ensured `custom_expires` was added as a property, things still didn't work because again nothing I could see in `SeedData` was adding the `expires_on` value we were supplying in `options`.

So my conclusion was in order to make this work

1. the fixture file must already have `custom_expires` and `expires_on` as properties. The existing ability to override any top level attribute in `SeedData` would handle updating `expires_on` to whatever we pass in via `options`. This makes the special `custom_expires` arg to `SeedData` unnecessary

2. the fixture file would need `custom_expires` to already be set. `SeedData` would then need additional logic that added the `expires_on` provided in options if it found `custom_expires` in the seed data. Again this makes the special `custom_expires` arg unnecessary

3. `SeedData` would need additional logic to add the `custom_expires` and `expires_on` properties to the seed data when the special `custom_expires` arg is set

4. the fixture file would need `expires_on` to already be set. You could then use the existing ability to override any top level attribute to set your dynamic `expires_on` value. You would no longer need the special `custom_expires` arg at all. But you would need to update the back-office API not to expect the `custom_expires` property

5. some super whizzy logic that did things depending on whether the fixture and/or the file had these properties set. Lets not go there!

Out of these I went for option 4. It does mean I've had to make a change to the API (see https://github.com/DEFRA/waste-carriers-back-office/pull/921). But as its the one option I felt meant I got to delete stuff rather than add, its the one I went for.