Add sleep to get next action to work
I don't know why the sleep is needed here. With the API updated and my previous changes I could see the link in the page but SitePrism kept throwing `cannot find css` error until I added a sleep.

To be honest, I'm sure there is a better more idomatic way to do this. But at this point I'm just trying to give confidence that the seed data changes both here and in the API work.